Create README.md
This project is a prototype of my personal portfolio. It includes the designing of a website with an Estate price predictor which aims at helping prospective estate buyers to know the available estates and their corresponding prices and also make purchases via the website. The other constituent of my personal portfolio is a Blog. The purpose of the blog is to disseminate rare information and other intriguing news around the world.
Through the creation of this portfolio, I have learnt the use of tools such as UI/Us, Figma, and also improved my understanding of Microsoft Excel, Word and Access. I have gained skills in the use of Python, Git and Figma. In addition, I have improved my typing and computing  skills and the overall process has made me a highly adaptable, open minded and resilient person
As a green leaf to such a project, I was faced with many challenges which included adapting to coding of various sorts and time constraint.